AAAClient - fixed project configuration
git-svn-id: http://hcm-lab.de/public/Horde3D/trunk@649 5ce291ac-9df0-446f-9e4f-d57731c4dda7

+++ Tools/GameEngine/Samples/AAAClient/main.cpp
@@ -92,7 +92,7 @@ int main()
 
 		//Send
 		//send( s, buffer_send, strlen(buffer_send), 0 );
-		sendto(s, buffer_send, strlen(buffer_send), 0, (struct sockaddr *)&addr, sizeof(addr));
+		sendto(s, buffer_send, (int)strlen(buffer_send), 0, (struct sockaddr *)&addr, sizeof(addr));
 
 		if(rc==SOCKET_ERROR)
 		{
